requirements.txt
Flask==2.2.2
keras==2.10.0
numpy==1.21.6
opencv_python==4.5.4.60
Pillow==9.2.0
tensorflow==2.10.0
